refactor and upd project to deploy and poublish package for web and nodejs environments
