+++ .gitignore
@@ -1,6 +1,4 @@
 /target
 **/*.rs.bk
 .vscode
-/js/node_modules
-/js/lib
-/js/target
+
