+++ js/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "doughnut-js"
-version = "0.1.1"
+version = "0.1.0"
 authors = ["The Root Network Team"]
 edition = "2021"
 description = "JS/Wasm bindings for the rust Doughnut codec"
