+++ js/build.sh
@@ -9,23 +9,20 @@ if ! [ -x "$(command -v jq)" ]; then
 fi
 
 # Clean previous packages
-if [ -d "pkg" ]; then
-    rm -rf pkg
+if [ -d "doughnut-web" ]; then
+    rm -rf doughnut-web
 fi
 
-if [ -d "pkg-node" ]; then
-    rm -rf pkg-node
+if [ -d "doughnut-nodejs" ]; then
+    rm -rf doughnut-nodejs
 fi
 
-PKG_NAME="doughnut"
+# build for web js target
+rustup run nightly wasm-pack build --target web --scope therootnetwork --out-name doughnut-web --release --out-dir doughnut-web
+# modify package.json for web
+jq '.name="@therootnetwork/doughnut-web"' doughnut-web/package.json > temp.json && mv temp.json doughnut-web/package.json
 
-# Build for both targets
-rustup run nightly wasm-pack build --target web --scope therootnetwork --out-name $PKG_NAME --release
-rustup run nightly wasm-pack build --target nodejs --scope therootnetwork --out-name $PKG_NAME --release --out-dir pkg-node
-
-# Merge nodejs & browser packages into `pkg/` directory
-cp "pkg-node/${PKG_NAME}.js" "pkg/${PKG_NAME}_main.js"
-sed "s/require[\(]'\.\/${PKG_NAME}/require\('\.\/${PKG_NAME}_main/" "pkg-node/${PKG_NAME}.js" > "pkg/${PKG_NAME}_bg.js"
-jq ".files += [\"${PKG_NAME}_bg.js\"]" pkg/package.json \
-  | jq ".main = \"${PKG_NAME}_main.js\"" > pkg/temp.json
-mv pkg/temp.json pkg/package.json
+# build for nodejs target
+rustup run nightly wasm-pack build --target nodejs --scope therootnetwork --out-name doughnut-nodejs --release --out-dir doughnut-nodejs
+# modify package.json for nodejs
+jq '.name="@therootnetwork/doughnut-nodejs"' doughnut-nodejs/package.json > temp.json && mv temp.json doughnut-nodejs/package.json
